refactoring and optimization
