+++ package/src/scripts/methods/clickCalendar.ts
@@ -7,6 +7,7 @@ import generateDate from '../helpers/generateDate';
 import mainMethod from './mainMethod';
 import handlerMultipleRanged from './handlerMultipleRanged';
 import getColumnID from '../helpers/getColumnID';
+import calendarInput from '../helpers/calendarInput';
 
 const clickCalendar = (self: IVanillaCalendar) => {
 	(self.HTMLElement as HTMLElement).addEventListener('click', (e) => {
@@ -118,20 +119,9 @@ const clickCalendar = (self: IVanillaCalendar) => {
 				if (self.actions.clickDay) self.actions.clickDay(e, self.selectedDates);
 
 				if (self.input && self.HTMLInputElement && self.HTMLElement && self.actions.changeToInput) {
-					const calendar = {
-						hide() {
-							(self.HTMLElement as HTMLElement).classList.add(self.CSSClasses.calendarHidden);
-						},
-						show() {
-							(self.HTMLElement as HTMLElement).classList.remove(self.CSSClasses.calendarHidden);
-						},
-						HTMLInputElement: self.HTMLInputElement,
-						HTMLElement: self.HTMLElement as HTMLDivElement,
-					};
-
 					self.actions.changeToInput(
 						e,
-						calendar,
+						calendarInput(self),
 						self.selectedDates,
 						self.selectedTime,
 						self.selectedHours,
